Adding debugging code to figure out uwsgi configuration
In multi-process mode, load balancing to many processes works perfectly,
but due to the many processes, there's no internal sharing of shared
data, leading to inconsistencies (most notably with the autocompletion
database). This is insanely annoying. In thread mode, however, the load
balancing did not work (properly?). Trying to find out systematically
what is wrong and solve this problem once and for all.